+++ CHAPTER6/キャラクタ/敵機/FN移動.py
@@ -2,8 +2,7 @@
 #┃キャラクタ：敵機：移動機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import  pyxel
-from    ._DB   import *
-from    ._ID   import 機体ID
+from    .DB import *
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
