--- CHAPTER6/キャラクタ/敵機/_ID.py
@@ -1,43 +1,0 @@
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃キャラクタ：敵機：ID
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from enum import IntEnum, auto
-
-class 機体ID(IntEnum):
-
-    # 特殊
-    機雷        = 0
-    ブロック    = auto()
-
-    # 戦闘機
-    戦闘機1     = auto()
-    戦闘機2     = auto()
-    戦闘機3     = auto()
-
-    # アイテム
-    アイテム    = auto()
-
-
-class アイテムID(IntEnum):
-
-    # 補給系
-    弾薬回復	    = 0
-    シールド回復	= auto()
-
-    # 永続効果
-    速度UP	        = auto()    # 0.5ずつ速くなる（上限あり）
-    発射_連射   	= auto()    # 永続効果
-    発射_貫通弾	    = auto()    # 永続効果
-    発射_弾数UP   	= auto()    # 永続効果、1ずつ増える（上限あり）
-
-    # 多方向同時発射
-    発射_上下方向   = auto()    # 発射_弾数UP と併用可
-    発射_左右方向   = auto()    # 発射_弾数UP と併用可
-    発射_4方向   	= auto()    # 発射_弾数UP と併用可
-    発射_8方向   	= auto()    # 発射_弾数UP は一時無効
-
-    # 敵からのダメージを緩和
-    ダメージ半減	= auto()    # ダメージが半減する
-    ダメージ無し	= auto()    # ノーダメージになる
-    ダメージ吸収    = auto()    # 敵の攻撃力をシールドに加える
-
