+++ CHAPTER6/キャラクタ/敵機/__init__.py
@@ -3,5 +3,4 @@
 #┃キャラクター・モジュール
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 from .main  import 本体 as 敵機出現
-from ._ID   import *
-from ._DB   import *
+from .DB   import *
