+++ CHAPTER6/キャラクタ/自機/DB.py
@@ -1,29 +1,41 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃キャラクター：標的
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from ._ID import *
+from    enum import IntEnum, auto
+
+class アイテムID(IntEnum):
+
+    # 補給系
+    弾薬回復	    = 0
+    シールド回復	= auto()
+
+    # 永続効果
+    速度UP	        = auto()    # 0.5ずつ速くなる（上限あり）
+    発射_連射   	= auto()    # 永続効果
+    発射_貫通弾	    = auto()    # 永続効果
+    発射_弾数UP   	= auto()    # 永続効果、1ずつ増える（上限あり）
+
+    # 多方向同時発射
+    発射_上下方向   = auto()    # 発射_弾数UP と併用可
+    発射_左右方向   = auto()    # 発射_弾数UP と併用可
+    発射_4方向   	= auto()    # 発射_弾数UP と併用可
+    発射_8方向   	= auto()    # 発射_弾数UP は一時無効
+
+    # 敵からのダメージを緩和
+    ダメージ半減	= auto()    # ダメージが半減する
+    ダメージ無し	= auto()    # ノーダメージになる
+    ダメージ吸収    = auto()    # 敵の攻撃力をシールドに加える
+
+class 効果ID(IntEnum):
+
+    追加速度     = 1
+    連射         = auto()
+    防御UP       = auto()
+    貫通弾       = auto()
+    発射数       = auto()
+    発射方向     = auto()
 
 class DB:
-
-    #[0] 得点
-    #[1] 移動
-    #   [0] 発速度Y      ： 0～
-    #   [1] 速度X        ： 0～(0:横移動なし)
-    #   [2] 方向反転間隔 ： 左右反転の間隔をfpsで指定(0:反転しない)
-    #[2] 発射
-    #   [0] 間隔： fpsで指定(0:発射しない)
-    #   [1] 弾数： 同時に発射する弾数(0:発射しない/1/2/4)
-    #   [2] 速度： 移動速度(0:発射しない)
-    #[3] 破壊力
-    #   [0] 弾  ： 弾の攻撃で与えるダメージ
-    #   [1] 衝突： 体当たりで与えるダメージ
-    敵機 = {
-            機体ID.戦闘機1 : ( 30, ( 1.2, 0.0,  0 ), ( 40, 2, 1.6 ), ( 5 ,3 ) ),
-            機体ID.戦闘機2 : ( 90, ( 1.0, 1.2, 30 ), ( 40, 1, 2.0 ), ( 4 ,2 ) ),
-            機体ID.戦闘機3 : ( 60, ( 0.8, 0.0,  0 ), ( 80, 4, 1.3 ), ( 5 ,5 ) ),
-            機体ID.機雷    : (  0, ( 0.3, 0.8, 20 ), (  0, 0, 0.0 ), ( 0 ,0 ) ),
-            機体ID.アイテム: (  0, ( 0.3, 0.8, 20 ), (  0, 0, 0.0 ), ( 0 ,0 ) ) }
-
     #[0] 出現情報 
     #   [0] 難易度     ： 難易度がこの数値異常で出現する 
     #   [1] 出現率     ： 難易度がこの数値異常で出現する
