+++ CHAPTER6/キャラクタ/自機/FN描画.py
@@ -2,7 +2,7 @@
 #┃キャラクタ：自機：描画機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import  pyxel
-from    ._ID    import 効果ID
+from    .DB import 効果ID
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
